+++ utility/scoring.hpp
@@ -6,7 +6,7 @@ using namespace std;
 /// @param gap Actual gap
 /// @param variance Expected gap
 /// @return The value of the mentioned function
-double normalDistribution (double gap, double variance) {
+long double normalDistribution (long double gap, long double variance) {
     return exp(-squared(gap) / (2 * squared(variance)));
 }
 
@@ -14,21 +14,21 @@ double normalDistribution (double gap, double variance) {
 namespace spatial {
     const double stdDeviation = 50;
 
-    double observationProbability (const Graph &G, const CandidatePoint &cp) {
-        double distanc = euclideDist(G.edges[cp.assocEdge].polyline, cp.gps), scalar = 1.0;
-        double expo = normalDistribution(distanc, stdDeviation);
+    long double observationProbability (const Graph &G, const CandidatePoint &cp) {
+        long double distanc = euclideDist(G.edges[cp.assocEdge].polyline, cp.gps), scalar = 1.0;
+        long double expo = normalDistribution(distanc, stdDeviation);
         return scalar * expo;
     }
 
-    double transmissionProbability (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
-        double shortestPath = shortestPathLength(G, from, to);
-        double euclid = euclideDist(from.gps, to.gps);
+    long double transmissionProbability (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
+        long double shortestPath = shortestPathLength(G, from, to);
+        long double euclid = euclideDist(from.gps, to.gps);
         // This formula is from Can Yang's paper since Yin Lou's formula left some problematic bugs
-        double probability = min(shortestPath, euclid) / max(shortestPath, euclid);
+        long double probability = min(shortestPath, euclid) / max(shortestPath, euclid);
         return probability;
     }
 
-    double score (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
+    long double score (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
         return observationProbability(G, to) * transmissionProbability(G, from, to);
     }
 };
@@ -38,9 +38,9 @@ namespace temporal {
     const double largeVariance = 30;
     const double smallVariance = 5;
 
-    double score (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
-        double timeLimit = fastTravelTime(G, from, to);
-        double actualTime = to.recordTime - from.recordTime;
+    long double score (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
+        long double timeLimit = fastTravelTime(G, from, to);
+        long double actualTime = to.recordTime - from.recordTime;
         if (actualTime < timeLimit)
             return normalDistribution(timeLimit - actualTime, largeVariance);
         return normalDistribution(actualTime - timeLimit, smallVariance);
@@ -52,6 +52,6 @@ namespace temporal {
 /// @param from Starting candidate point
 /// @param to Ending candidate point
 /// @return The scoring according to the paper
-double SPFunction (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
+long double SPFunction (const Graph &G, const CandidatePoint &from, const CandidatePoint &to) {
     return spatial::score(G, from, to) * temporal::score(G, from, to); // combining spatial & temporal analysis function
 }
